refactor account, account_linear
