+++ README.md
@@ -40,6 +40,8 @@ The following API endpoints have been implemented
 - `/v2/private/order/create` Place Active Order
 - `/v2/private/order/list` Get Active Order
 - `/v2/private/order/cancel` Cancel Active Order
+- `/v2/private/position/list` My Position
+- `/v2/private/position/leverage/save` Set Leverage
 
 #### Wallet Data Endpoints
